requirements.txt
streamlit==1.28.1
pandas==2.0.3
gspread==5.12.0
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
requests==2.31.0
